% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_drift.R
\name{alt_drift}
\alias{alt_drift}
\title{Drift Analysis for ALT test.}
\usage{
alt_drift(
  admin,
  testcode,
  mode,
  status,
  RSnum,
  mock,
  dnu = "",
  flagDist = "",
  dropuin = "",
  dropseq = "",
  dropuin_anchor = "",
  direction = "normal",
  get_plot = T,
  anchor_var = "linking_status",
  iter = 0
)
}
\arguments{
\item{admin}{String.}

\item{testcode}{String.}

\item{mode}{P or E.}

\item{status}{OP or FT.}

\item{RSnum}{RS1-3}

\item{mock}{T or F.}

\item{dnu}{Do-NOT-USE item in UIN in Calibration.}

\item{flagDist}{Items that have flag and dropped in Calibration}

\item{dropuin}{Items dropped in Calibration}

\item{dropseq}{IRTLOC dropped in Calibration}

\item{dropuin_anchor}{Item that to be dropped from anchor set.}

\item{direction}{Normal or Backward}

\item{get_plot}{T for F}

\item{anchor_var}{variable that used to identify anchor set. "preequating" or "linking_status"}

\item{iter}{nth iteration.Can be blank.}
}
\value{
irt_file: Item parameter file that contains equated pars and base pars.

mydrift_table: A table of drift items

drifted_items: List of drifted items

anchor_set: A table of anchor items.
}
\description{
Drift Analysis for ALT test.
}
