% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_d2check.R
\name{anchor_d2check}
\alias{anchor_d2check}
\title{compute D2 and WRMSD for drift analysis.}
\usage{
anchor_d2check(ne_par, ol_par, con_path, out_path, equate = T, get_plot = T)
}
\arguments{
\item{ne_par}{unscaled parameters}

\item{ol_par}{anchor parameters}

\item{con_path}{input for STUIRT.}

\item{out_path}{output from STUIRT.}

\item{equate}{T or F, whether to scale the pars in ne_par file.}

\item{get_plot}{T or F, whether to plot ICCs.}
}
\value{
m1m2: Stocking-Lord constants.

anch_prms: anchor item parameters.

ICC_plot: generate a pdf plot if get_plot is True.
}
\description{
compute D2 and WRMSD for drift analysis.
}
