% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2trait.R
\name{d2trait}
\alias{d2trait}
\title{Convert UIN with Ds to Trait UIN.}
\usage{
d2trait(uin_d, testcode)
}
\arguments{
\item{uin_d}{String.Trait UIN with Ds.}

\item{testcode}{String.}
}
\value{
uin_trait: Trait UIN with Long Strings.
}
\description{
Convert UIN with Ds to Trait UIN.
}
