% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqSTUIRT_4drift.R
\name{eqSTUIRT_4drift}
\alias{eqSTUIRT_4drift}
\title{Run equating analysis and evaluate stableness of the anchor set.}
\usage{
eqSTUIRT_4drift(
  admin,
  testcode,
  mode,
  status,
  RSnum,
  mock,
  purpose = "normal",
  direction = "normal",
  anchor_var = "PREEQUATING",
  dropuin = "",
  dropseq = "",
  dnu = "",
  dropuin_anchor = "",
  drop_passid = "",
  forcein_uin = "",
  forcein_passid = "",
  iter = "",
  get_plot = T
)
}
\arguments{
\item{admin}{String. Administration.}

\item{testcode}{String.}

\item{mode}{E or P}

\item{RSnum}{String. RS1-3.}

\item{mock}{T or F.}

\item{purpose}{normal or maintain.}

\item{direction}{normal or backward.}

\item{anchor_var}{variable used for anchor. "PREEQUATING" or "LINKING_STATUS".}

\item{dropuin}{UIN that dropped in calibration.}

\item{dropseq}{IRTLOC that dropped in calibration.}

\item{dnu}{Do-NOT-USE items in UIN.}

\item{dropuin_anchor}{Item UIN that to be dropped from anchor set.}

\item{drop_passid}{PassageID that to be dropped from anchor set.}

\item{forcein_uin}{Item UIN that is forced to be in anchor set.}

\item{forcein_passid}{PassageID that is forced to be in anchor set.}

\item{iter}{number of iteration cycle.}

\item{get_plot}{T or F. Whether to get plot.}
}
\value{
anch_file: Parameter file about anchor items.

passageDrift: Parameter file for passage drift analysis.

drifted: List of items that are drifted.

BP_cover: Blueprint coverage of the anchor set.

irt_file: Scaled and unscaled parameter file.

internalirt: Item parameter file for internal use.

m1m2: Stocking_Lord constants from STUIRT.

base_vs_freeirt_summary: Summary statistics about base and unscaled parameters.

plot_reivew: plots.
}
\description{
Run equating analysis and evaluate stableness of the anchor set.
}
