% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2_int_adj.R
\name{G2_int_adj}
\alias{G2_int_adj}
\title{Compute true points for each points for an item. This function is used in conjuction with model fit.}
\usage{
G2_int_adj(theta.vec.short, Ncount, item, freeirt = freeirt)
}
\arguments{
\item{theta.vec.short}{A vector of thetas.}

\item{Ncount}{A vector of counts.}

\item{item}{1,2,3...}

\item{freeirt}{A table contains irt parameters.}
}
\value{
N_P_item: A list that contains counts of each points for each theta values for the specified item.
}
\description{
Compute true points for each points for an item. This function is used in conjuction with model fit.
}
