% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ab.R
\name{get_ab}
\alias{get_ab}
\title{get initial values for STUIRT.}
\usage{
get_ab(ol_par, ne_par)
}
\arguments{
\item{ol_par}{base parameters.}

\item{ne_par}{unscaled free parameters.}
}
\value{
A and B: initial values for STURIT.
}
\description{
get initial values for STUIRT.
}
