% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctt.R
\name{get_ctt}
\alias{get_ctt}
\title{Compute CTT statistics.}
\usage{
get_ctt(irt_par, idm)
}
\arguments{
\item{irt_par}{A table of item parameters.}

\item{idm}{IDM.}
}
\value{
myctt: A table of CTT statistics.
}
\description{
Compute CTT statistics.
}
