% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_d2.R
\name{get_d2}
\alias{get_d2}
\title{get D2 based on base and equated pars.}
\usage{
get_d2(
  model,
  theta.vec,
  D,
  base_a,
  base_b,
  base_c,
  base_steps,
  equated_a,
  equated_b,
  equated_c,
  equated_steps,
  uin,
  get_plot = T
)
}
\arguments{
\item{model}{"rasch","2PL","3PL", or "GPC".}

\item{theta.vec}{A vector of thetas.}

\item{D}{1 or 1.702.}

\item{base_a}{Base A parameter.}

\item{base_b}{Base B parameter.}

\item{base_c}{Base C parameter.}

\item{base_steps}{A vector of Base steps.}

\item{equated_a}{Equated A parameter .}

\item{equated_b}{Equated B parameter.}

\item{equated_c}{Equated C parameter.}

\item{equated_steps}{A vector of equated steps.}

\item{uin}{item UIN.}

\item{get_plot}{T or F.}
}
\value{
d2, wrmsd, and drift_plot.
}
\description{
get D2 based on base and equated pars.
}
