% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eprs.R
\name{get_eprs}
\alias{get_eprs}
\title{Generate score tables.}
\usage{
get_eprs(
  admin,
  testcode,
  mode,
  status = "OP",
  battery,
  RSnum,
  mock,
  pre_post = "post",
  what = c("pl3b", "semb", "ssb", "pl4b", "ssb", "irtb", "aberrant"),
  irt_add = ""
)
}
\arguments{
\item{admin}{String. Administration.}

\item{testcode}{String.}

\item{mode}{E or P}

\item{status}{OP or FT.}

\item{battery}{1 or 0.}

\item{RSnum}{String. RS1-3.}

\item{mock}{T or F.}

\item{pre_post}{PRE or POST.}

\item{what}{Types of score table. E.g., "pl3b","semb","ssb","pl4b","ssb","irtb","aberrant".}

\item{irt_add}{A table contains parameters to add to bank file.}
}
\value{
Score table that specified in the `what` argument.
}
\description{
Generate score tables.
}
