% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eprs_heading.R
\name{get_eprs_heading}
\alias{get_eprs_heading}
\title{Generate special headings for score tables.}
\usage{
get_eprs_heading(admin, what, battery)
}
\arguments{
\item{admin}{String.}

\item{what}{Type of score tables.}

\item{battery}{1 or 0.}
}
\value{
heading: headings that specified in `what` argument.
}
\description{
Generate special headings for score tables.
}
