% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{get path for files. Use this function in conjunction of `newadmin_setup`, which will create directories for the files.}
\usage{
get_path(
  admin,
  testcode,
  mode,
  status = NA,
  mock,
  RSnum,
  pre_post = NA,
  purpose = NA,
  iter = ""
)
}
\arguments{
\item{admin}{String.}

\item{testcode}{String.}

\item{mode}{P or E.}

\item{status}{OP or FT.}

\item{mock}{T or F.}

\item{RSnum}{RS1-3.}

\item{pre_post}{pre or post.}

\item{purpose}{normal or maintain.}

\item{iter}{integer, e.g., 1,2,3... iterations}
}
\value{
list of paths.
}
\description{
get path for files. Use this function in conjunction of `newadmin_setup`, which will create directories for the files.
}
