% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rsss.R
\name{get_rsss}
\alias{get_rsss}
\title{Generate raw-to-scale score file.}
\usage{
get_rsss(
  admin,
  testcode,
  mode,
  status = "OP",
  RSnum,
  mock,
  pre_post,
  irt_add = ""
)
}
\arguments{
\item{admin}{String. Administration.}

\item{testcode}{String.}

\item{mode}{E or P}

\item{status}{OP or FT.}

\item{RSnum}{String. RS1-3.}

\item{mock}{T or F.}

\item{pre_post}{PRE or POST.}

\item{irt_add}{A table contains parameters to add to bank file.}
}
\value{
rsss: Raw to Scale Score file.

pl3b: Raw to Scale Score file for reporting categories.
}
\description{
Generate raw-to-scale score file.
}
