% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_cal.R
\name{irt_cal}
\alias{irt_cal}
\title{Run IRTpro/Winsteps calibration.}
\usage{
irt_cal(
  admin,
  testcode,
  mode,
  status,
  RSnum,
  mock,
  dnu = "",
  dropuin = "",
  dropseq = "",
  flagDist = ""
)
}
\arguments{
\item{admin}{String.}

\item{testcode}{String.}

\item{mode}{P or E.}

\item{status}{OP or FT.}

\item{RSnum}{RS1-3}

\item{mock}{T or F.}

\item{dnu}{Do-NOT-USE item in UIN.}

\item{dropuin}{Items that to drop in Calibration}

\item{dropseq}{IRTLOC that to drop in Calibration}

\item{flagDist}{Items that have flag and to drop in Calibration}
}
\value{
freeirt: Freely calibrated IRT parameters

convergence: Convergence message.
}
\description{
Run IRTpro/Winsteps calibration.
}
