% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtpro_fun.R
\name{irtpro_fun}
\alias{irtpro_fun}
\title{Run IRTpro calibration.}
\usage{
irtpro_fun(
  admin,
  testcode,
  mode,
  status,
  RSnum,
  mock,
  dnu = "",
  dropuin = "",
  dropseq = "",
  flagDist = ""
)
}
\arguments{
\item{admin}{String.}

\item{testcode}{String.}

\item{mode}{P or E.}

\item{status}{OP or FT.}

\item{RSnum}{RS1-3}

\item{mock}{T or F.}

\item{dnu}{Do-NOT-USE item in UIN.}

\item{dropuin}{Items that to drop in Calibration}

\item{dropseq}{IRTLOC that to drop in Calibration}

\item{flagDist}{Items that have flag and to drop in Calibration}
}
\value{
freeirt: Freely calibrated IRT parameters

convergence: Convergence message from IRTpro.

idm_cal: IDM that is calibrated in IRTpro.
}
\description{
Run IRTpro calibration.
}
