% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modfit_fun.R
\name{modfit_fun}
\alias{modfit_fun}
\title{Compute Q1 and G2 for model fit.}
\usage{
modfit_fun(admin, testcode, mode, status, RSnum, mock)
}
\arguments{
\item{admin}{String.}

\item{testcode}{String.}

\item{mode}{P or E.}

\item{status}{OP or FT.}

\item{RSnum}{RS1-3.}

\item{mock}{T or F.}
}
\value{
fit_tab: A table with fit statistics

fit_plot: A list of fit plots.
}
\description{
Compute Q1 and G2 for model fit.
}
