% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newadmin_setup.R
\name{newadmin_setup}
\alias{newadmin_setup}
\title{Set up directories for new administration.}
\usage{
newadmin_setup(admin_year)
}
\arguments{
\item{admin_year}{administration year. E.g., "Spring19", "Fall20"}
}
\description{
Set up directories for new administration.
}
