% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_irtpro.R
\name{out_irtpro}
\alias{out_irtpro}
\title{Extract IRTpro results.}
\usage{
out_irtpro(out_file_path)
}
\arguments{
\item{out_file_path}{Output from IRTpro.}
}
\value{
output: Extraction from IRTpro.

convergence: Convergence message from IRTpro.
}
\description{
Extract IRTpro results.
}
