% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_winsteps.R
\name{out_winsteps}
\alias{out_winsteps}
\title{Extract Winsteps results.}
\usage{
out_winsteps(out_file_path)
}
\arguments{
\item{out_file_path}{Output from Winsteps}
}
\value{
out_table Extraction from Winsteps

convergence: Convergence message from Winsteps.
}
\description{
Extract Winsteps results.
}
