% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_sas.R
\name{round_sas}
\alias{round_sas}
\title{function to round up at 5 to mimic SAS results}
\usage{
round_sas(x, n)
}
\arguments{
\item{x}{number.}

\item{n}{integer. Which decimal place that want to round to.}
}
\value{
A rounded number
}
\description{
function to round up at 5 to mimic SAS results
}
