% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{function to round up at 5}
\usage{
round_up(x, n)
}
\arguments{
\item{x}{number.}

\item{n}{integer. Which decimal place that want to round to.}
}
\value{
A rounded number
}
\description{
function to round up at 5
}
