% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait2d.R
\name{trait2d}
\alias{trait2d}
\title{Convert Trait UIN to UIN with Ds.}
\usage{
trait2d(trait_uin)
}
\arguments{
\item{trait_uin}{Trait UIN with Long Strings.}
}
\value{
uin_d: UIN with Ds.
}
\description{
Convert Trait UIN to UIN with Ds.
}
