% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_dist.R
\name{um_dist}
\alias{um_dist}
\title{Compute distribution abnormal.}
\usage{
um_dist(score_vec, score_cat, max_point, N_TOTAL)
}
\arguments{
\item{score_vec}{A vector of scores.}

\item{score_cat}{Score category. E.g., A binary item would have score_cat = 2 (1,0).}

\item{max_point}{Max point of an item. E.g.,  A binary item would have max_point = 1.}

\item{N_TOTAL}{Total N counts.}
}
\value{
um_dist: "U" means "U-shaped" distribution. "M" means missing score category .
}
\description{
Compute distribution abnormal.
}
