% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsteps_fun.R
\name{winsteps_fun}
\alias{winsteps_fun}
\title{Run Winsteps calibration.}
\usage{
winsteps_fun(
  admin,
  testcode,
  mode,
  status,
  RSnum,
  mock,
  dnu = "",
  dropuin = "",
  dropseq = "",
  flagDist = ""
)
}
\arguments{
\item{admin}{String.}

\item{testcode}{String.}

\item{mode}{P or E.}

\item{status}{OP or FT.}

\item{RSnum}{RS1-3}

\item{mock}{T or F.}

\item{dnu}{Do-NOT-USE item in UIN.}

\item{dropuin}{Items that to drop in Calibration}

\item{dropseq}{IRTLOC that to drop in Calibration}

\item{flagDist}{Items that have flag and to drop in Calibration}
}
\value{
freeirt: Freely calibrated IRT parameters

convergence: Convergence message from Winsteps.

to_consider: Items that consider to drop.
}
\description{
Run Winsteps calibration.
}
